Cleanup trakt injection
