+++ core/trakt-api/implementation/src/androidMain/kotlin/com/thomaskioko/trakt/service/implementation/TraktAuthInterceptor.kt
@@ -1,7 +1,6 @@
 package com.thomaskioko.trakt.service.implementation
 
 import android.content.SharedPreferences
-import com.thomaskioko.tvmaniac.util.model.Configs
 import io.ktor.http.HttpHeaders
 import me.tatarka.inject.annotations.Inject
 import net.openid.appauth.AuthState
@@ -10,7 +9,6 @@ import okhttp3.Response
 
 @Inject
 class TraktAuthInterceptor(
-    private val configs: Configs,
     private val authPrefs: SharedPreferences,
 ) : Interceptor {
 
@@ -22,8 +20,6 @@ class TraktAuthInterceptor(
         request = request.newBuilder()
             .addHeader(HttpHeaders.ContentType, "application/json")
             .addHeader(HttpHeaders.Authorization, "Bearer ${authState?.accessToken}")
-            .addHeader("trakt-api-version", "2")
-            .addHeader("trakt-api-key", configs.traktClientId)
             .build()
         return chain.proceed(request)
     }
